+++ docs/history.md
@@ -8,7 +8,7 @@ Important changes are emphasized.
 
 - **Breaking:** Drop support for Python 3.8
 - Add support for Python 3.13
-- Remove `cffi` as a runtime dependency
+- Remove all runtime dependencies (`cffi` & `asn1crypto`)
 - Add `COINCURVE_VENDOR_CFFI` environment variable to control vendoring of the `_cffi_backend` module
 - Minor performance improvement by removing use of formatted string constants
 - Upgrade [libsecp256k1][] to version 0.6.0